tests/wpt/tests/css/css-anchor-position/anchor-position-multicol-003.html

<!DOCTYPE html>
<title>Tests anchors on out-of-flow boxes</title>
<link rel="help" href="https://drafts.csswg.org/css-anchor-1/#determining">
<link rel="help" href="https://drafts.csswg.org/css-anchor-1/#propdef-anchor-name">
<link rel="help" href="https://drafts.csswg.org/css-anchor-1/#anchor-size">
<link rel="author" href="mailto:kojii@chromium.org">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/check-layout-th.js"></script>
<script src="support/test-common.js"></script>
<style>
.relpos {
  position: relative;
}
.columns {
  column-count: 2;
  column-fill: auto;
  column-gap: 10px;
  column-width: 100px;
  width: 210px;
  height: 50px;
}
.anchor1 {
  anchor-name: --a1;
  position: absolute;
  width: 10px;
  height: 30px;
  background: orange;
}
.target {
  position: absolute;
  height: anchor-size(--a1 height);
  background: lime;
}
</style>
<body onload="checkLayoutForAnchorPos('.target')">
  <div class="spacer" style="height: 10px"></div>
  <div class="relpos">
    <div class="columns">
      <div class="relpos">
        <div class="spacer" style="height: 30px"></div>
        <div class="anchor1"></div>
        <div class="target" data-expected-height=50></div>
      </div>
      <div class="target" data-expected-height=50></div>
    </div>
    <div class="target" data-expected-height=50></div>
  </div>

  <div class="spacer" style="height: 10px"></div>
  <div class="relpos">
    <div class="columns">
      <div class="spacer" style="height: 10px"></div>
      <div class="relpos">
        <div class="spacer" style="height: 10px"></div>
        <div class="relpos">
          <div class="spacer" style="height: 10px"></div>
          <div class="anchor1"></div>
          <div class="target" data-expected-height=50></div>
        </div>
        <div class="target" data-expected-height=50></div>
      </div>
      <div class="target" data-expected-height=50></div>
    </div>
    <div class="target" data-expected-height=50></div>
  </div>
</body>
