LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/anchor-position-multicol-003-expected.txt


FAIL .target 1 assert_equals:
<div class="target" data-expected-height="50"></div>
height expected 50 but got 30
PASS .target 2
PASS .target 3
FAIL .target 4 assert_equals:
<div class="target" data-expected-height="50"></div>
height expected 50 but got 30
FAIL .target 5 assert_equals:
<div class="target" data-expected-height="50"></div>
height expected 50 but got 30
PASS .target 6
PASS .target 7
